Change:  stored procedure InsertTransfer

+++ T-Store.DB/StoredProcedure/Transaction_Insert.sql
@@ -1,5 +1,5 @@
 ﻿create procedure [dbo].[Transaction_Insert]
-	@AccountId int,
+	@AccountId bigint,
 	@TransactionType tinyint,
 	@Amount decimal (11,4),
 	@Currency smallint
